ai_service/requirements.txt
cachetools
chromadb
fastapi
httpx
langchain
langchain-chroma
langchain-openai
pydantic-ai
pydantic-settings
python-dotenv
uvicorn[standard]
